Raise the floor: lint the code, test the charts, gate both in CI
An audit rather than a feature. Measured first, then fixed what the numbers
showed.

Dead code the linter found that reading had not. The important one:
VOL_INDEX_MAP was defined in both data/prices.py and data/providers.py after
the provider refactor. The two copies happened to agree, which is exactly how
that bug waits — add a ticker to one and they diverge silently. Now one
definition, re-exported. Plus an unused local in backtest, american, svi and
three in plots.

zip(..., strict=True) wherever two sequences are paired that the surrounding
code already assumes are aligned. A silent truncation becomes a loud error,
which for a chart means a visible failure instead of a quietly wrong picture.
Verified against live SPY and every chart in both languages, since strict= is a
behaviour change and the tests alone would not have proven it safe.

Chart coverage went from 34% to 94%. Twenty chart functions existed and five
were exercised; the rest had been verified by looking at them once, which
catches a bad layout and misses everything after. The new file renders every
one, in both languages, plus the degenerate cases — and immediately found a
singular y-axis on a single-expiry heatmap.

utils.py went from 66% to 100%. Retries and the cache sit under every market
data call and every second pipeline run, and were among the least tested code
here — the wrong way round, since a cache serving stale data fails quietly and
looks like a data problem.

CI now runs ruff and fails under 80% coverage, so none of this drifts back.

156 tests (was 109), 86% coverage, lint clean.

+++ volsurface/analytics/skew.py
@@ -49,7 +49,7 @@ def skew_metrics(surface, delta: float = 0.25) -> pd.DataFrame:
         curv = float((surface.iv(eps, T) - 2 * atm + surface.iv(-eps, T)) / eps ** 2)
 
         rows.append({
-            "days": int(round(float(T) * CALENDAR_DAYS)),
+            "days": round(float(T) * CALENDAR_DAYS),
             "T": float(T),
             "forward": float(surface.forward(T)),
             "atm_iv": atm,
@@ -157,7 +157,7 @@ def detect_anomalies(surface, cfg: AnalyticsConfig | None = None,
     for v in cal.get("violations", []):
         findings.append({
             "category": "calendar_arbitrage",
-            "tenor_days": int(round(v["T_far"] * CALENDAR_DAYS)),
+            "tenor_days": round(v["T_far"] * CALENDAR_DAYS),
             "metric": "total variance decreasing in T",
             "value": -v["max_depth_total_var"],
             "benchmark": 0.0,
@@ -177,7 +177,7 @@ def detect_anomalies(surface, cfg: AnalyticsConfig | None = None,
             continue
         findings.append({
             "category": "butterfly_arbitrage",
-            "tenor_days": int(round(s.T * CALENDAR_DAYS)),
+            "tenor_days": round(s.T * CALENDAR_DAYS),
             "metric": "min Durrleman g(k)",
             "value": float(getattr(s, "min_durrleman_g", np.nan)),
             "benchmark": 0.0,
